pyproject.toml
[project]
name = "info-hub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.121.0",
    "httpx>=0.28.1",
    "pgvector>=0.4.1",
    "pydantic-settings>=2.11.0",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "ruff>=0.14.3",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.38.0",
]

[tool.ruff]
line-length = 100
lint.select = ["E", "F", "I", "UP", "B"]
lint.ignore = ["E501"]  # optional: ignore line length
